Core, WASM: move generated-loader patches into tsdown config
Run the napi and wasm post-build patches from a `build:done` tsdown hook
instead of standalone `scripts/patch-*.ts`, matching how the helpers package
does its codegen in config. The hook fires once per output format, so a flag
spans both fires. Drops patching the napi `index.js` (it is the bundler input,
not shipped). Published output is unchanged.
