+++ takumi-napi/package.json
@@ -63,7 +63,7 @@
   },
   "scripts": {
     "build": "napi build --release --no-const-enum --platform --esm --strip",
-    "build:js": "tsdown && bun scripts/patch-dist-loader.ts",
+    "build:js": "tsdown",
     "prepublishOnly": "bun artifacts && napi prepublish -t npm --no-gh-release",
     "artifacts": "napi create-npm-dirs && napi artifacts && napi version",
     "bench": "bun tests/bench/index.tsx",
